Add comprehensive tests for HypurrFiVault contract
- Create MockHypurrFiOracle for testing
- Fix REFERRAL_CODE type from uint256 to uint16 in HypurrFiVault
- Add HypurrFiVault.test.ts with 78 comprehensive tests covering:
  * Deployment and initialization
  * Deposit functionality and leverage loop execution
  * Withdrawal and position unwinding
  * Redeem operations
  * Health factor calculations and monitoring
  * Rebalancing mechanisms
  * ERC-4626 compliance
  * Edge cases (first depositor, large deposits/withdrawals, multiple users)
  * Security features (reentrancy protection, access control)
  * Owner functions (pause, emergency deleverage, parameter updates)
  * View functions
  * Integration tests with realistic user scenarios

All tests passing (78/78)

+++ packages/hardhat/contracts/HypurrFiVault.sol
@@ -77,7 +77,7 @@ contract HypurrFiVault is ERC4626, ReentrancyGuard, Ownable {
     // Constants
     uint256 private constant BASIS_POINTS = 10000;
     uint256 private constant VARIABLE_INTEREST_RATE_MODE = 2;
-    uint256 private constant REFERRAL_CODE = 0;
+    uint16 private constant REFERRAL_CODE = 0;
 
     // ==================== Modifiers ====================
